Relocate Opt In classes
